+++ trakt/__init__.py
@@ -4,6 +4,6 @@ try:
 except ImportError:
     pass
 
-version_info = (2, 0, 3)
+version_info = (2, 1, 0)
 __author__ = 'Jon Nappi'
 __version__ = '.'.join([str(i) for i in version_info])